build(ex): update a new ex items organization
