+++ CMakeLists.txt
@@ -1,6 +1,8 @@
 CMAKE_MINIMUM_REQUIRED(VERSION 3.15.0)
 PROJECT(run VERSION 0.1.0)
 
+set(CMAKE_CXX_STANDARD 17)
+
 # enable_testing()
 
 # INCLUDE(cmake/utils.cmake)
@@ -10,5 +12,6 @@ PROJECT(run VERSION 0.1.0)
 # ADD_SUBDIRECTORY(utils)
 # add_subdirectory(test)
 add_subdirectory(playground)
+add_subdirectory(ex)
 
 set(CMAKE_CXX_STANDARD 17)
