--- ex/main.cpp
@@ -1,1 +1,0 @@
-int main() { return 0; }
