--- ex/producer_consumer.cpp
@@ -1,16 +1,0 @@
-#include <condition_variable>
-#include <iostream>
-#include <mutex>
-#include <string>
-#include <thread>
-
-#include "./factory_method.hpp"
-
-class Storage {};
-
-class Producer {
- public:
- private:
-};
-
-class Consumer {};
