+++ ex/ring_buffer/main.cpp
@@ -1,0 +1,6 @@
+#include <iostream>
+
+int main() {
+  std::cout << "Hello World!" << std::endl;
+  return 0;
+}
