+++ ex/ring_buffer/ring_buffer.h
@@ -1,0 +1,1 @@
+#pragma once
